Rename tests/dummy.py to tests/utils.py
